Upload sample output from "sc.sql"
This sample output shows what may be expected when running the "sc.sql" script.

Of course, specific parameter settings and timing values are merely examples and any actual results should not be compared to these.  Remember that the "sc.sql" script is intended for comparisons on your own systems, between two databases on different platforms, or between changes on the same database.